requirements.txt
PyYAML>=6.0
typing>=3.7.4.1
requests>=2.27.1
argparse>=1.4.0
pandas>=1.3.5
parse>=1.15.0